Clean-up code using RatingScore
Part of #1070

+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/person/ai/task/meta/ExamineBodyMeta.java
@@ -57,7 +57,7 @@ public class ExamineBodyMeta  extends MetaTask implements SettlementMetaTask {
 	private static final double DEFAULT_SCORE = 500D;
 
 	// Extra score for every day body not examined
-	private static final int SOL_SCORE = 50;
+	private static final double SOL_SCORE = 50;
 
 	private static MedicalManager medicalManager;
